MapR 11281 (contd.): Drop database cascade fails if any tables have indexes created as separate tables

+++ metastore/src/java/org/apache/hadoop/hive/metastore/HiveMetaStoreClient.java
@@ -497,7 +497,7 @@ public class HiveMetaStoreClient implements IMetaStoreClient {
           // already deleted
         }
 
-        if (table != null)
+        if (table != null && !isIndexTable(table))
           dropTable(table, deleteData, false);
       }
     }
